requirements.txt
numpy
datetime
re
os
opencv-python
click
pdf2image
python-poppler
imutils
pytorch-pretrained-bert
pyenchant